backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11
Pillow>=10.0
celery>=5.3
django-celery-beat>=2.6
redis>=5.0
scikit-learn>=1.3
pandas>=2.1
numpy>=1.26
gunicorn>=21.2
psycopg2-binary>=2.9
